feat(leaderboard): Display full name from registration data
The leaderboard previously only displayed Advent of Code usernames, which can be difficult to associate with actual participants.

This change introduces a feature to show the full name of each participant. A registration CSV file is now fetched and parsed on application load to map AoC usernames to full names.

A "Full Name" column has been added to the leaderboard table to display this information, making the standings more user-friendly and recognizable. Users not found in the registration file will be marked as "Not Registered".